+++ i18n/trubar-config.yaml
@@ -5,9 +5,9 @@ languages:
   si:
     name: Slovenščina
     international-name: Slovenian
-    auto-import: from orangecanvas.utils.localization.si import plsi, plsi_sz, z_besedo  # pylint: disable=wrong-import-order
+    auto-import: from orangecanvas.localization.si import plsi, plsi_sz, z_besedo  # pylint: disable=wrong-import-order
 auto-import: |2
-  from orangecanvas.utils.localization import Translator  # pylint: disable=wrong-import-order
+  from orangecanvas.localization import Translator  # pylint: disable=wrong-import-order
   _tr = Translator("orangecanvas", "biolab.si", "Orange")
   del Translator
 encoding: "utf-8"
